only save patches of mutation results
Instead of saving the entire cache for each optimistic response, only
save the minimum data for the optimistic response. We also need to pass
into the optimistic reducer the results of the APOLLO_MUTIATION_INIT
action. I left a comment in the code to make that more clear. Also,
corrected the logic for filtering out the optimistic response once the
real response is applied.