Add redirects for existing React Apollo 2.0 docs URLs.
These redirects maintain the URLs in existence before the React Apollo 2.1
docs overhaul.

This includes both Netlify redirects for redirects that should be handled by
the server (specifically, those that don't include anchor tags on the "from"
side) and also a JavaScript-powered `redirects` directive in the `_config.yml`
which utilizes a mechanism in the theme.

Ref: https://github.com/apollographql/apollo-client/pull/3097