fix(fetch-more): prevent fetch more from breaking after a non-graphql error (#2906)
* fix(fetch-more): prevent fetch more from breaking after a non-graphql error

When using fetch more, the original query is set as the fetchMoreForQueryId.
In the old solution, when a non-graphql error occurred, both the query to fetch more and the original query were marked as erroneous. Thus both of them were removed from the store. This left the application in a weird state with null pointers and other unwanted side effects.

In this new solution we mark the fetch-more query as erroneous, but we set the original query as completed. Because the original query has already been fetched, and should not be invalidated/removed when the fetch-more query fails.

* fix fetch more tests

* add changelog