Fix typescript issue with `defaultOptions` in constructor
When the new `QueryBaseOptions` and `QueryOptions` interfaces
were created in
https://github.com/apollographql/apollo-client/commit/f443620a7b86418361502e22b29e2a9643229865,
`ModifiableWatchQueryOptions` was replaced by `QueryOptions`
when checking the `query` `defaultOptions` used in the
Apollo Client constructor. Unfortunately, this introduced a
small type check bug since the `query` property in `QueryOptions`
is a mandatory property, whereas `ModifiableWatchQueryOptions`
did not include this property. This commit replaces the use of
`QueryOptions` with `QueryBaseOptions` when type checking
`query` in `defaultOptions`, since `QueryBaseOptions` doesn't
have a `query` property.

Fixes https://github.com/apollographql/apollo-client/issues/3583.