docs: Use example.com rather than my-app.com.
Example.com is an IANA-recognized[[0]] test domain intended for using in
examples, whereas my-app.com is an actual business.

I only came across this while checking for broken links.  Luckily, the
broken link checker that I use ignores example.com (as I'd _almost_ expect it
to, because of the IANA "Example" designation.)

[0]: https://www.iana.org/domains/reserved