Snapshot observableQuery.lastResult to defend against mutation.
Should help with #3992.

Though deep cloning and isEqual testing are obviously more expensive than
relying on === equality, the reality is that lastResult may have been
modified since it was previously recorded, so === equality is useless for
determining isDifferentResult. A defensive copy is the only way to make
sure we get this right.