Avoid async/await to reduce bundle size.
Part of #4224.

This change reduces the minified+gzip'd bundle size of apollo-client from
10385 bytes to 9771 bytes, a ~6% decrease, according to this command:

  npx terser -m -c < lib/bundle.umd.js | gzip | wc -c

As long as we're only building one bundle, and it has to work in the 15%
minority of browsers that do not natively support async/await, we should
avoid wasting precious bundle size on generator runtime code, especially
since we had only one await expression in the entire package (so the
runtime code was nowhere close to paying for itself).