Un-revert "Improve (and shorten) query polling implementation. (#4243)"
This reverts commit 29168c156dd1d1de37c948f521d912af3475586c.

Now that we have a uniform interface for terminating ApolloClient
instances (#4336), there should be no need for any external code to access
the QueryScheduler abstraction, which this commit removes.

We should wait to merge and release this change until after
https://github.com/apollographql/react-apollo/pull/2741 has been merged
and released, so that we don't break older versions of MockedProvider.