Remove the "Docs" CI job (#4362)
We have a build process setup via Netlify to automatically
checkout and build this proejct's docs, along with Netlify's
github check setup so we know when a docs build/deploy fails
for any reason. This failure is reported back on PR's just like
the "Docs" CI job is, so there isn't really any reason to
duplicate this work.