Avoid including Observable multiplex operator in production bundle.
Multiplexing the Observable returned by getObservableFromLink in the
startGraphQLSubscription method seemed necessary because we have tests to
verify that subscription results/errors are delivered to all subscribers,
with titles like 'should multiplex subscriptions'.

However, upon closer inspection, the reason those tests were failing is
that MockSubscriptionLink clobbers previous observers on this line:
https://github.com/apollographql/apollo-client/blob/acfb98ebd78ba1e6289b54b4dd89f8acd1ddddc1/packages/apollo-client/src/__mocks__/mockLinks.ts#L117

What MockSubscriptionLink is doing is a bug that would need to be fixed if
it occurred in any actual application, presumably by keeping track of all
observers and properly distributing the mocked results.

Fortunately, we already have a handy utility for supporting an unlimited
number of observers, without subscribing to the underlying Observable more
than once. That's right: we can simply move the multiplex call into
MockSubscriptionLink, which means it will no longer be included in in the
production bundle, which saves 57 minified+gzip bytes.