Revert "concrete-versions-for-react-apollo"
This reverts commit 056a540c3eec1aa37e88dc9d8d1501eaaf121a08, which was
intended to be a temporary fix to enable running

  npm install ../apollo-client/packages/apollo-client

in my local checkout of the react-apollo repository.

Installing a package from a relative path often doesn't work well with
file:... dependency versions in package.json. However, we definitely want
to use file:... dependencies among the apollo-client/packages/*, so that
we're depending on the current local versions of those packages, rather
than depending on the last published versions.