Use ID+field dependencies for result caching, rather than just ID.
Until now, the result caching system registered dependencies on specific
entity objects by ID, which meant that changes to any of the fields of an
entity in the cache would invalidate every query result that previously
involved that entity, even if some of those queries did not use any of the
fields that changed.

While this system was logically correct, it resulted in extensive and
unnecessary recomputation of cached query results. Perhaps most
dramatically, any changes to the fields of the ROOT_QUERY object would
necessitate the recomputation of every query that involved any root query
fields, because those queries depended on the ROOT_QUERY ID rather than
the specific fields they consumed.

With this commit, the StoreReader reads from the EntityCache using a new
method called getFieldValue, which registers dependencies on the
combination of the entity ID and the name of the requested field, so that
cached query results will remain valid as long as the specific fields they
used have not changed.

Furthermore, the EntityCache#set method has been replaced by a method
called merge, which intelligently updates the normalized data using a set
of new fields, automatically invalidating dependencies for any fields
whose values are modified by the merge.

For backwards compatibility, if a consumer chooses to continue using the
EntityCache#get(id) method instead of getFieldValue(id, fieldName),
ID-based dependencies will be registered, as before.