Revert "Throw if Policies#fragmentMatches receives a falsy __typename."
Fixes #5709 by reverting commit 329fa1041ddf59d805330ebfaffec140796a3e1b.

The removed test was invalid because a fragment with a type condition can
never match an object without a __typename field in Apollo Client 3.0.
The invariant was related to this test because the invariant failed
earlier, effectively hiding the true misbehavior of the test.