Prevent `@client` fields from being passed into the Link chain
Apollo Client 2.x allowed `@client` fields to be passed into the
`link` chain if `resolvers` were not set in the constructor. This
allowed `@client` fields to be passed into Links like
`apollo-link-state`. Apollo Client 3 enforces that `@client` fields
are local only, meaning they are no longer passed into the `link`
chain, under any circumstances. We're making this change as we're
slowly starting to deprecate the Local State API (and local
resolvers) in favor of the new Type Policies API. Since we want
people to be able to use the Type Policies API (and `@client`)
without local resolvers, we don't want to force them to always
define `resolvers: {}` in their `ApolloClient` constructor, to
prevent `@client` fields from being passed into the `link` chain.

If anyone wants to preserve this behavior in AC 3, they can
consider using their own directive (instead of `@client`), and
using that directive in the `link` chain.