Add graphql/execution/execute to external list in rollup.config.js.
This prevents the @apollo/client/link/schema CommonJS bundle (which is
used by Node.js) from accidentally pulling in the entire dependency tree
of graphql/execution/execute.

This duplication not only increased the @apollo/client/link/schema CJS
bundle size from 1.31kB to a whopping 22.6kB, but also caused instanceof
to stop working reliably for types like GraphQLSchema, because there might
be more than one definition of that constructor in play, and instanceof is
sensitive to the exact constructor function you pass as the right-hand
argument.

Should fix #6621, allowing us to revert #6622. Thanks to @stolinski for
surfacing this issue.