Pass QueryInfo into ObservableQuery constructor.
This gives the ObservableQuery direct access to its corresponding
QueryInfo object, so the ObservableQuery doesn't have to keep asking the
QueryManager for the QueryInfo corresponding to this.queryId.

Important invariants: given ObservableQuery oq and QueryManager qm, it
should always be the case that qm.queries.get(oq.queryId) ===
oq.queryInfo, and oq.queryInfo.observableQuery === oq.