Fix optimistic mutation tests, again.
This essentially undoes commit 9b3aa0d448fe014be630faf1daf7ba1deeb52ef2,
which was introduced as part of PR #6221. I remember thinking, as I made
that commit, "This change shouldn't be necessary, but it seems harmless
enough." As it turns out, changing these tests was not necessary, but was
instead a symptom of overreacting to the delivery of optimistic mutation
results. I'm happy to be putting things back the way they were.