docs: make error policy 'none' description clearer
I was a bit confused by the text here because it sounded like it was saying that GraphQL errors are treated literally the same as network errors (ie, they would actually show up on `error.networkError` somehow). The description above the table was more explicit, but it didn't actually say what the default policy was, so I wasn't sure that the clearer description actually described `none`.