Make refetch results overwrite existing data.
This is how I would ideally like to fix #7491, but I'm worried it could be
a breaking change for any application code that relies on refetch results
getting merged with existing data, rather than replacing it (which is what
refetch was originally meant to do, since AC2).

I will follow this commit with one that makes this the overwrite behavior
opt-in (and thus backwards compatible).